+++ src/pub/app.js
@@ -18,7 +18,6 @@ import Login from './login.js';
 import Overview from './overview.js';
 import Config from './config.js';
 import Settings from './settings.js';
-import Tracer from './tracer.js';
 import Proxy_add from './proxy_add.js';
 import Whitelist_ips from './whitelist_ips.js';
 import {Logs, Dock_logs} from './logs.js';
@@ -207,7 +206,6 @@ const Page = ()=>
             <Route path="/howto/:option?/:suboption?" exact component={Howto}/>
             <Route path="/logs" exact component={Logs}/>
             <Route path="/proxy_tester" exact component={Proxy_tester}/>
-            <Route path="/tracer" exact component={Tracer}/>
             <Route path="/config" exact component={Config}/>
             <Route path="/settings" exact component={Settings}/>
             <Route path="/" component={Overview}/>