+++ util/lpm_util.js
@@ -60,6 +60,7 @@ E.init_args = args=>{
         dir: 'Path to the directory with database and configuration files',
         status: 'Show proxy manager processes current status',
         'show-logs': 'Show logs of the currently running LPM instance',
+        'gen-cert': 'Generate cert',
     };
     const usage = ['Usage:\n  $0 [options] config1 config2 ...'];
     if (process.env.DOCKER)